Small schema changes

+++ database/migrations/2018_02_23_224218_create_reservations_table.php
@@ -19,7 +19,7 @@ class CreateReservationsTable extends Migration
             $table->date('end_date');
             $table->timestamp('when_logged');
             $table->boolean('is_active');
-            $table->bigInteger('vehicle_id')->unsigned();
+            $table->integer('vehicle_id')->unsigned();
             $table->foreign('vehicle_id')->references('id')->on('vehicles');
         });
     }
